Refactor factory to internal package.
