+++ plugins/org.komodo.relational/src/org/komodo/relational/model/internal/ParameterImpl.java
@@ -16,7 +16,6 @@ import org.komodo.relational.RelationalConstants.Direction;
 import org.komodo.relational.RelationalConstants.Nullable;
 import org.komodo.relational.model.Parameter;
 import org.komodo.relational.model.Procedure;
-import org.komodo.relational.model.RelationalModelFactory;
 import org.komodo.relational.model.StatementOption;
 import org.komodo.spi.KException;
 import org.komodo.spi.repository.KomodoObject;
