+++ tests/org.komodo.relational.test/src/org/komodo/relational/model/internal/IndexImplTest.java
@@ -18,7 +18,6 @@ import org.junit.Test;
 import org.komodo.relational.RelationalModelTest;
 import org.komodo.relational.model.Column;
 import org.komodo.relational.model.Index;
-import org.komodo.relational.model.RelationalModelFactory;
 import org.komodo.relational.model.Table;
 import org.komodo.relational.model.TableConstraint;
 import org.modeshape.sequencer.ddl.dialect.teiid.TeiidDdlLexicon;
