+++ tests/org.komodo.relational.test/src/org/komodo/relational/model/internal/PrimaryKeyImplTest.java
@@ -13,7 +13,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.komodo.relational.RelationalModelTest;
 import org.komodo.relational.model.PrimaryKey;
-import org.komodo.relational.model.RelationalModelFactory;
 import org.komodo.relational.model.Table;
 import org.komodo.relational.model.TableConstraint;
