+++ tests/org.komodo.relational.test/src/org/komodo/relational/model/internal/TableImplTest.java
@@ -22,7 +22,6 @@ import org.komodo.relational.model.Column;
 import org.komodo.relational.model.ForeignKey;
 import org.komodo.relational.model.Index;
 import org.komodo.relational.model.PrimaryKey;
-import org.komodo.relational.model.RelationalModelFactory;
 import org.komodo.relational.model.SchemaElement.SchemaElementType;
 import org.komodo.relational.model.StatementOption;
 import org.komodo.relational.model.Table;
